+++ fr3d/database/DbMgr.py
@@ -8,7 +8,7 @@ import pymysql
 from pymysql.connections import Connection
 from pymysql.cursors import DictCursor
 
-from constants.DDatabase import DDatabase
+from fr3d.constants.DDatabase import DDatabase
 
 class DbMgr: